requirements.txt
ccxt
python-telegram-bot
matplotlib
numpy
mysql-connector-python